Renamed cli-engine to cli
